+++ services/registry/src/validation/validators/schemaValidator.ts
@@ -1,11 +1,11 @@
 import { filter, map, take } from "rxjs/operators";
 import {makeExecutableSchema} from "graphql-tools";
 import { AgogosObjectConfig, SchemaConfig } from "../../sync/object-types";
-import { makeGqlDocumentFromGqlSources } from "../../sync/sync-schemas";
 import gqlObjects$, { AggObjsByName } from "../../sync/sync-service";
+import { stitchSchemas } from "../../graphql/stitch";
 
 export const validateSchemas = (schemas: AggObjsByName) => {
-    const stitchedSchema = makeGqlDocumentFromGqlSources(schemas);
+    const stitchedSchema = stitchSchemas(schemas);
     makeExecutableSchema({
         typeDefs: stitchedSchema,
     });
